requirements.txt
aniso8601==9.0.1
argcomplete==1.10.0
attrs==20.3.0
backcall==0.2.0
beautifulsoup4==4.8.0
Bootstrap-Flask==1.5.1
camelot-py==0.8.2
cffi==1.14.5
chardet==3.0.4
click==7.1.2
cryptography==3.4.6
cssselect==1.1.0
cycler==0.10.0
dateparser==1.0.0
decorator==4.4.2
distro==1.5.0
docx2txt==0.8
dominate==2.6.0
EbookLib==0.17.1
et-xmlfile==1.0.1
extract-msg==0.23.1
Flask==1.1.2
flask-marshmallow==0.14.0
flask-restplus==0.13.0
Flask-SQLAlchemy==2.4.4
Flask-Uploads==0.2.1
Flask-WTF==0.14.3
fpdf==1.7.2
gunicorn==20.0.4
IMAPClient==2.1.0
invoice2data==0.3.5
ipykernel==5.5.0
ipython==7.21.0
ipython-genutils==0.2.0
itsdangerous==1.1.0
jdcal==1.4.1
jedi==0.18.0
Jinja2==2.11.3
jsonschema==3.2.0
jupyter-client==6.1.11
jupyter-core==4.7.1
kiwisolver==1.3.1
lxml==4.6.2
MarkupSafe==1.1.1
marshmallow==3.10.0
marshmallow-sqlalchemy==0.24.2
matplotlib==3.3.4
numpy==1.20.1
olefile==0.46
opencv-python==4.5.1.48
openpyxl==3.0.6
pandas==1.2.3
parso==0.8.1
pdfkit==0.6.1
pdfminer.six==20181108
pdfquery==0.4.3
pdfrw==0.4
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.1.2
prompt-toolkit==3.0.16
ptyprocess==0.7.0
pycparser==2.20
pycryptodome==3.10.1
Pygments==2.8.1
PyMuPDF==1.18.9
pyparsing==2.4.7
PyPDF2==1.26.0
pyquery==1.4.3
pyrsistent==0.17.3
pytesseract==0.3.7
python-dateutil==2.8.1
python-pptx==0.6.18
pytz==2021.1
PyYAML==5.4.1
pyzmq==22.0.3
regex==2020.11.13
reportlab==3.5.63
roman==3.3
six==1.12.0
sortedcontainers==2.3.0
soupsieve==2.2
SpeechRecognition==3.8.1
SQLAlchemy==1.3.23
tabula-py==2.2.0
textract==1.6.3
tornado==6.1
traitlets==5.0.5
tzlocal==1.5.1
Unidecode==1.2.0
visitor==0.1.3
wcwidth==0.2.5
Werkzeug==1.0.1
WTForms==2.3.3
xlrd==1.2.0
XlsxWriter==1.3.7
